split to one more class
